KNN algorithm users
This is a K nearest neighbors classifier algorithm (https://en.wikipedia.org/wiki/K-nearest_neighbors_algorithm) in which the engine predict the classification into a class and depends on neighbors number and metrics (20 and Euclidean distance in our case)
In our case accuracy it is quite low. It would be usual because the more data and historical (less zeros) the more accuracy. 
In this case y=users